changed int() to float() once
no significant changes.
the problem seems to be, when the converter library is called a part is called to times, once with a dict, which the converter can handle and once with a pandas Series, which the converter cannot handle

+++ savetojson.py
@@ -108,7 +108,7 @@ def read_year_and_budget(input_list, year):
         else:
             years_dict[str(items)] = {"selected": "", "CO2 limit": co2[index].item()}"""
         # the String format has resulted in an error in combination withe .item() function; now, it somehow works
-        years_dict[str(items)] = {"selected": "", "CO2 limit": int(co2[index])}
+        years_dict[str(items)] = {"selected": "", "CO2 limit": float(co2[index])}
         index = index + 1
 
     return years_dict
